Redirecting input to stdin from files now works for Python on Windows
Stdin through console overflows it's buffer when writing large number
of characters, so as an alternative, the input is generated and saved
into a file first, and then is redirected to the stdin of the program
file.

Currently the input runs only on Windows due to a new cmd window being
spawned by the process. The redirect operator does not work purely
using exec.Command() because it's not a command argument.

Also, output from stdout is not displayed anymore because it causes
stdout buffer overflow for large outputs.

Output from stderr is taken and displayed instead in case it contains
useful debugging information.